edemco-backend-python/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.0
outcome==1.3.0.post0
pandas==2.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
pywin32==311
requests==2.32.5
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
watchdog==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
xlrd==2.0.1
xlwings==0.31.7
